// packages/story-engine/src/types/StoryMetrics.ts
export interface StoryMetrics {
  totalChanges: number;
  totalCommits: number;
  totalContributors: number;
  complexityScore: number;
  riskScore: number;
  impactScore: number;
  confidenceScore: number;
  dependencyCount: number;
  decisionCount: number;
  bugCount: number;
  refactorCount: number;
  sessionCount: number;
}